Add VersionAdapter to support 'replace-ocean-biomes' on 1.18+
The benefit of VersionAdapter over VersionUtils is that each adapter
implementation is built against its corresponding Paper/Spigot version,
meaning it's easier to catch errors and less reflection has to be used.

However, VersionUtils is tightly coupled with the rest of the codebase
which makes it hard to migrate some methods without substantial
refactoring. That'll be something to look at in the future.